+++ src/main/java/lotto/domain/User.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import lotto.utils.Money;
 
 public class User {
