+++ src/main/java/lotto/utils/NumberException.java
@@ -1,12 +1,10 @@
-package lotto.exception;
+package lotto.utils;
 
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
-
-import lotto.domain.Money;
 
 public class NumberException {
