+++ src/main/java/lotto/view/GameInput.java
@@ -1,7 +1,7 @@
 package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
-import lotto.exception.NumberException;
+import lotto.utils.NumberException;
 import lotto.domain.User;
 
 public class GameInput {
